将size_t强制转换为int

+++ Plugins/Stock/Common.cpp
@@ -164,8 +164,8 @@ std::vector<std::string> CCommon::split(std::string str, std::string pattern)
     std::string::size_type pos;
     std::vector<std::string> result;
     str += pattern;//扩展字符串以方便操作
-    size_t size = str.size();
-    for (size_t i = 0; i < size; i++)
+    int size = static_cast<int>(str.size());
+    for (int i = 0; i < size; i++)
     {
         pos = str.find(pattern, i);
         if (pos < size)
